move load data scripts to dedicated folder. separated all sims, renamed files, small pep8 changes
