+++ synthesizer/load_data/load_camels.py
@@ -5,6 +5,7 @@ import numpy as np
 from astropy.cosmology import FlatLambdaCDM
 
 from .particle.galaxy import Galaxy
+from utils import get_len
 
 
 def _load_CAMELS(
@@ -241,7 +242,10 @@ def load_CAMELS_Astrid(
         form_time = hf["PartType4/GFM_StellarFormationTime"][:]
         coods = hf["PartType4/Coordinates"][:]
         masses = hf["PartType4/Masses"][:]
-        imasses = np.ones(len(masses)) * 0.00155  # TODO: update with correct scaling
+
+        # TODO: update with correct scaling
+        imasses = np.ones(len(masses)) * 0.00155
+
         _metals = hf["PartType4/GFM_Metals"][:]
         metallicity = hf["PartType4/GFM_Metallicity"][:]
 
@@ -375,89 +379,3 @@ def load_CAMELS_SIMBA(
         star_forming=star_forming,
         dtm=dtm,
     )
-
-def load_FLARES(f, region, tag):
-    """
-    Load FLARES galaxies from a FLARES master file
-
-    Args:
-        f (string):
-            master file location
-        region (string):
-            FLARES region to load
-        tag (string):
-            snapshot tag to load
-
-    Returns:
-        galaxies (object):
-            `ParticleGalaxy` object containing stars
-    """
-
-    with h5py.File(f, "r") as hf:
-        lens = hf[f"{region}/{tag}/Galaxy/S_Length"][:]
-        ages = hf[f"{region}/{tag}/Particle/S_Age"][:]  # Gyr
-        coods = hf[f"{region}/{tag}/Particle/S_Coordinates"][:].T
-        mass = hf[f"{region}/{tag}/Particle/S_Mass"][:]  # 1e10 Msol
-        imass = hf[f"{region}/{tag}/Particle/S_MassInitial"][:]  # 1e10 Msol
-        # metals = hf[f'{region}/{tag}/Particle/S_Z'][:]
-        metals = hf[f"{region}/{tag}/Particle/S_Z_smooth"][:]
-        s_oxygen = hf[f"{region}/{tag}/Particle/S_Abundance_Oxygen"][:]
-        s_hydrogen = hf[f"{region}/{tag}/Particle/S_Abundance_Hydrogen"][:]
-        # ids = hf[f'{region}/{tag}/Particle/S_ID'][:]
-        # index = hf[f'{region}/{tag}/Particle/S_Index'][:]
-        # hf[f'{pre}/S_Vel']
-
-    ages = ages * 1e9  # yr
-    mass = mass * 1e10  # Msol
-    imass = imass * 1e10  # Msol
-
-    begin, end = get_len(lens)
-
-    galaxies = [None] * len(begin)
-    for i, (b, e) in enumerate(zip(begin, end)):
-        galaxies[i] = Galaxy()
-        galaxies[i].load_stars(
-            mass[b:e],
-            ages[b:e],
-            metals[b:e],
-            s_oxygen=s_oxygen[b:e],
-            s_hydrogen=s_hydrogen[b:e],
-            coordinates=coods[b:e, :],
-        )
-
-    return galaxies
-
-
-def get_len(Length):
-    """
-    Find the beginning and ending indices from a length array
-
-    Args:
-        Length (array):
-            array of number of particles
-
-    Returns:
-        begin (array):
-            beginning indices
-        end (array):
-            ending indices
-    """
-
-    begin = np.zeros(len(Length), dtype=np.int64)
-    end = np.zeros(len(Length), dtype=np.int64)
-    begin[1:] = np.cumsum(Length)[:-1]
-    end = np.cumsum(Length)
-    return begin, end
-
-
-def main():
-    # e.g.
-    # '/cosma7/data/dp004/dc-love2/codes/flares/data/FLARES_30_sp_info.hdf5'
-    _f = sys.argv[1]
-    tag = sys.argv[2]  # '/010_z005p000/'
-    load_FLARES(_f, tag)
-    return None
-
-
-if __name__ == "__main__":
-    main()
