+++ synthesizer/load_data/load_sc-sam.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 from synthesizer.parametric.sfzh import BinnedSFZH
 from synthesizer.parametric.galaxy import Galaxy as ParametricGalaxy
 from synthesizer.particle.stars import Stars
@@ -18,7 +19,7 @@ def load_SCSAM(fname, method, grid=None, verbose=False):
         fname (str): 
             SC-SAM sfhist_*-*.dat file to be read
         method (str): 
-            'particle', 'parametric_NNI' or 'paremteric_RGI',
+            'particle', 'parametric_NNI' or 'parametric_RGI',
             depending on how you wish to model your SFZH.
             - 'particle' treats each age-Z bin as a particle
             - 'parametric_NNI' uses scipy's nearest ND interpolator to
@@ -44,7 +45,7 @@ def load_SCSAM(fname, method, grid=None, verbose=False):
     lines = sfhist.readlines()
 
     # Set up halo index, birth halo ID, redshift and galaxy object lists
-    halo_ind_list  = []
+    halo_ind_list = []
     birthhalo_id_list = []
     redshift_list = []
     galaxies = []
@@ -60,7 +61,8 @@ def load_SCSAM(fname, method, grid=None, verbose=False):
         if count == 0:
             Z_len, age_len = [int(i) for i in line.split()]
             if verbose:
-                print(f'There are {Z_len} metallicity bins and {age_len} age bins.')
+                print((f'There are {Z_len} metallicity bins ',
+                       f'and {age_len} age bins.'))
 
         # Get metallicity bins
         if count == 1:
@@ -71,14 +73,14 @@ def load_SCSAM(fname, method, grid=None, verbose=False):
             if len(Z_lst) != Z_len:
                 print('Wrong number of Z bins.')
                 break
-            Z_sun = 0.02 # Solar metallicity
-            Z_lst = 10**np.array(Z_lst) * Z_sun # Unitless
+            Z_sun = 0.02  # Solar metallicity
+            Z_lst = 10**np.array(Z_lst) * Z_sun  # Unitless
             if verbose:
                 print(f'Z_lst (unitless): {Z_lst}')
 
         # Get age bins
         if count == 2:
-            age_lst = [float(i) for i in line.split()] # Gyr
+            age_lst = [float(i) for i in line.split()]  # Gyr
             if verbose:
                 print(f'age_lst: {age_lst}')
             # Check that this agrees with the expected grid structure
@@ -118,13 +120,14 @@ def load_SCSAM(fname, method, grid=None, verbose=False):
 
             # Create galaxy object
             if method=='particle':
-                galaxy = _load_SCSAM_particle_galaxy(SFH, age_lst, Z_lst, verbose=verbose)
+                galaxy = _load_SCSAM_particle_galaxy(
+                    SFH, age_lst, Z_lst, verbose=verbose)
             elif method=='parametric_NNI':
-                galaxy = _load_SCSAM_parametric_galaxy(SFH, age_lst,
-                                                       Z_lst, 'NNI', grid, verbose=verbose)
+                galaxy = _load_SCSAM_parametric_galaxy(
+                    SFH, age_lst, Z_lst, 'NNI', grid, verbose=verbose)
             elif method=='parametric_RGI':
-                galaxy = _load_SCSAM_parametric_galaxy(SFH, age_lst,
-                                                       Z_lst, 'RGI', grid, verbose=verbose)
+                galaxy = _load_SCSAM_parametric_galaxy(
+                    SFH, age_lst, Z_lst, 'RGI', grid, verbose=verbose)
 
             # Append to list of galaxy objects
             galaxies.append(galaxy)
@@ -146,9 +149,9 @@ def _load_SCSAM_particle_galaxy(SFH, age_lst, Z_lst, verbose=False):
     """
 
     # Initialise arrays for storing particle information
-    p_imass = [] # initial mass
-    p_age = [] # age
-    p_Z = [] # metallicity
+    p_imass = []  # initial mass
+    p_age = []  # age
+    p_Z = []  # metallicity
 
     # Get length of arrays
     age_len = len(age_lst)
@@ -159,19 +162,19 @@ def _load_SCSAM_particle_galaxy(SFH, age_lst, Z_lst, verbose=False):
         print('Iterating through grid...')
     for age_ind in range(age_len):
         for Z_ind in range(Z_len):
-            if SFH[age_ind][Z_ind]==0:
+            if SFH[age_ind][Z_ind] == 0:
                 continue
             else:
-                p_imass.append(SFH[age_ind][Z_ind]) # 10^9 Msun
-                p_age.append(age_lst[age_ind]) # Gyr
-                p_Z.append(Z_lst[Z_ind]) # unitless
+                p_imass.append(SFH[age_ind][Z_ind])  # 10^9 Msun
+                p_age.append(age_lst[age_ind])  # Gyr
+                p_Z.append(Z_lst[Z_ind])  # unitless
 
     # Convert units
     if verbose:
         print('Converting units...')
-    p_imass = np.array(p_imass) * 10**9 # Msun
-    p_age = np.array(p_age) * 10**9 # yr
-    p_Z = p_Z # unitless
+    p_imass = np.array(p_imass) * 10**9  # Msun
+    p_age = np.array(p_age) * 10**9  # yr
+    p_Z = p_Z  # unitless
 
     if verbose:
         print('Generating SED...')
@@ -188,8 +191,14 @@ def _load_SCSAM_particle_galaxy(SFH, age_lst, Z_lst, verbose=False):
     return particle_galaxy
 
 
-def _load_SCSAM_parametric_galaxy(SFH, age_lst, Z_lst, method, grid, verbose=False):
-
+def _load_SCSAM_parametric_galaxy(
+        SFH,
+        age_lst,
+        Z_lst,
+        method,
+        grid,
+        verbose=False
+):
     """
     Obtain galaxy SED using the parametric method.
     This is done by interpolating the grid.
@@ -207,18 +216,18 @@ def _load_SCSAM_parametric_galaxy(SFH, age_lst, Z_lst, method, grid, verbose=Fal
 
     # This the grid that we want to interpolate to
     new_age = 10**grid.log10age # yr
-    new_Z = np.log10(grid.metallicity) # log10Z
+    new_Z = np.log10(grid.metallicity)  # log10Z
 
     # This is the old grid, to be interpolated
-    old_age = np.array(age_lst) * 10**9 # yr
-    old_Z = np.log10(Z_lst) # log10Z
+    old_age = np.array(age_lst) * 10**9  # yr
+    old_Z = np.log10(Z_lst)  # log10Z
 
     # Convert SFH units
-    SFH = np.array(SFH) * 10**9 # Msun
-    sum_SFH = np.log10(np.sum(SFH))
+    SFH = np.array(SFH) * 10**9  # Msun
+    # sum_SFH = np.log10(np.sum(SFH))
 
     # Using regular grid interpolator
-    if method=='RGI':
+    if method == 'RGI':
         # Get coords for new grid
         new_X, new_Y = np.meshgrid(new_age, new_Z, indexing='ij')
         # Set up the old grid for interpolation
@@ -231,7 +240,7 @@ def _load_SCSAM_parametric_galaxy(SFH, age_lst, Z_lst, method, grid, verbose=Fal
         new_SFH[np.isnan(new_SFH)] = 0.
 
     # Using nearest ND interpolator
-    if method=='NNI':
+    if method == 'NNI':
         # Get coords for new grid
         new_X, new_Y = np.meshgrid(new_age, new_Z, indexing='ij')
         # Get list of coordinates of old grid
@@ -249,7 +258,7 @@ def _load_SCSAM_parametric_galaxy(SFH, age_lst, Z_lst, method, grid, verbose=Fal
     # Create Binned SFZH object
     sfzh = BinnedSFZH(log10ages=grid.log10age, metallicities=grid.metallicity,
                       sfzh=new_SFH)
-    
+
     # Create galaxy object
     parametric_galaxy = ParametricGalaxy(sfzh)
